Code reuse and jest
misc
