+++ mentors/jacob/2022-12-14/wordplay-only-u-03.test.js
@@ -1,0 +1,10 @@
+const {hasUniqueVowel, filterWords} = require('./wordplay-only-u-03.js');
+
+// https://jestjs.io/docs/getting-started
+test('hasUniqueVowel', () => {
+  expect(hasUniqueVowel('JILL', 'I')).toBe(true);
+  expect(hasUniqueVowel('JIILL', 'I')).toBe(true);
+  expect(hasUniqueVowel('JIOLL', 'I')).toBe(false);
+  expect(hasUniqueVowel('JOLL', 'I')).toBe(false);
+  expect(hasUniqueVowel('JLL', 'I')).toBe(false);
+});
